[#4] Ability to install as python package from git (#1)
