+++ .gitignore
@@ -10,3 +10,4 @@
 ehthumbs.db
 Thumbs.db
 
+.vscode
